3_resources/jax-basics/hklstm/src/requirements.txt
black~=22.0
flake8>=3.7.9, <5.0
ipython>=7.31.1, <8.0
isort~=5.0
jupyter~=1.0
jupyterlab_server>=2.11.1, <2.16.0
jupyterlab~=3.0
kedro~=0.18.4
kedro-telemetry~=0.2.0
nbstripout~=0.4
pytest-cov~=3.0
pytest-mock>=1.7.1, <2.0
pytest~=6.2
dm-env>=1.2
dm-tree>=0.1.1
packaging>=20.9
tensorflow-datasets>=4.2.0
tensorflow>=2.4.0
dm-haiku
optax
pandas
plotnine
